requirements.txt
Flask==2.0.1
gunicorn==19.7.1
Jinja2==3.0.1
flasgger==0.9.5
psycopg2-binary=2.9.3